Properly parse expiration time
The golang oauth2 library's `Token` doesn't have an "expires_in"
field. Instead oauth2 internally parses the expiration into a future
unix timestamp. So before this patch, the tokens that oauth2device
was producing didn't properly capture the expiration time. This patch
fixes it by copying the logic from oauth2, which is unfortunately
not exposed in a way that we can directly use.

Change-Id: Idec2632c5d4cb272d9d2e8e0099972ad9dda9c3d